feat: add prettier, eslint, typescript, etc
