+++ jest.config.js
@@ -1,0 +1,4 @@
+module.exports = {
+    testEnvironment: 'node',
+    preset: 'ts-jest',
+  };
