+++ package.json
@@ -1,19 +1,40 @@
 {
-    "name": "ws_task",
+    "name": "battleship",
     "version": "1.0.0",
-    "description": "Template for the WebSocket remote control task",
+    "description": "",
     "main": "index.js",
-    "type": "module",
     "scripts": {
-        "start": "node ./index.js",
-        "start:dev": "nodemon ./index.js",
-        "test": "echo \"Error: no test specified\" && exit 1"
+        "start": "node ./index.ts",
+        "start:dev": "nodemon ./index.ts",
+        "start:prod": "tcs && node dist/index.js",
+        "test": "jest --passWithNoTests",
+        "format": "prettier --write \"src/**/*.ts\"",
+        "lint": "eslint \"src/**/*.ts\" --fix"
     },
-    "author": "Andrei Auchynnikau",
-    "keywords": [],
+    "author": "Dmitry Strizhakov",
+    "keywords": [
+        "ws",
+        "websockets",
+        "battleship"
+    ],
     "license": "ISC",
     "devDependencies": {
-        "nodemon": "^2.0.16"
+        "@types/jest": "^29.5.12",
+        "@types/node": "^20.11.19",
+        "@types/ws": "^8.5.10",
+        "@typescript-eslint/eslint-plugin": "^7.0.1",
+        "@typescript-eslint/parser": "^7.0.1",
+        "dotenv": "^16.4.4",
+        "eslint": "^8.56.0",
+        "eslint-config-prettier": "^9.1.0",
+        "eslint-plugin-prettier": "^5.1.3",
+        "eslint-prettier-config": "^1.0.1",
+        "jest": "^29.7.0",
+        "nodemon": "^2.0.22",
+        "prettier": "^3.2.5",
+        "ts-jest": "^29.1.2",
+        "ts-node": "^10.9.2",
+        "typescript": "^5.3.3"
     },
     "dependencies": {
         "ws": "^8.8.0"
