+++ tsconfig.json
@@ -1,0 +1,21 @@
+{
+    "compilerOptions": {
+      "module": "CommonJS",
+      "moduleResolution": "Node",
+      "target": "ES2020",
+      "sourceMap": true,
+      "outDir": "dist",
+      "baseUrl": "src",
+      "forceConsistentCasingInFileNames": true,
+      "allowSyntheticDefaultImports": true,
+      "esModuleInterop": true,
+      "lib": ["ESNext"],
+      "noUncheckedIndexedAccess": true,
+      "strictNullChecks": true,
+      "removeComments": true,
+      "resolveJsonModule": true,
+      "skipLibCheck": true,
+      "strict": true,
+      "strictPropertyInitialization": false,
+    },
+  }
